Move some utils logic from WgpuCtx
