+++ compositor_render/src/render_loop.rs
@@ -27,7 +27,7 @@ pub(super) fn populate_inputs(
     ctx.wgpu_ctx.queue.submit([]);
 
     for (node, input_textures) in scene.inputs.values() {
-        ctx.wgpu_ctx.yuv_to_rgba_converter.convert(
+        ctx.wgpu_ctx.format.convert_yuv_to_rgba(
             ctx.wgpu_ctx,
             (input_textures.yuv_textures(), input_textures.bind_group()),
             &node.output.rgba_texture(),
@@ -42,7 +42,7 @@ pub(super) fn read_outputs(
 ) -> HashMap<OutputId, Arc<Frame>> {
     let mut pending_downloads = Vec::with_capacity(scene.outputs.len());
     for (node_id, (node, output)) in &scene.outputs {
-        ctx.wgpu_ctx.rgba_to_yuv_converter.convert(
+        ctx.wgpu_ctx.format.convert_rgba_to_yuv(
             ctx.wgpu_ctx,
             (&node.output.rgba_texture(), &node.output.bind_group()),
             output.yuv_textures(),
