+++ compositor_render/src/renderer/format/yuv_to_rgba.wgsl
